refactor: move CantonClient and related logic to a new module
